Use u32 for bandwidth monitor

+++ shared/src/connection/bandwidth_monitor.rs
@@ -2,7 +2,7 @@ use std::time::Duration;
 
 pub struct BandwidthMonitor {
     time_queue: TimeQueue<usize>,
-    total_bytes: u16,
+    total_bytes: u32,
     to_kbps_factor: f32,
 }
 
@@ -18,7 +18,7 @@ impl BandwidthMonitor {
     pub fn record_packet(&mut self, bytes: usize) {
         self.clear_expired_packets();
 
-        self.total_bytes += bytes as u16;
+        self.total_bytes += bytes as u32;
         self.time_queue.add_item(bytes);
     }
 
@@ -30,7 +30,7 @@ impl BandwidthMonitor {
 
     fn clear_expired_packets(&mut self) {
         while let Some(bytes) = self.time_queue.pop_item() {
-            self.total_bytes -= bytes as u16;
+            self.total_bytes -= bytes as u32;
         }
     }
 }
